+++ lua/ui/lobby/changelog.lua
@@ -65,7 +65,13 @@ changelog = {
             "- 'Set ranked options' button works again",
             "- Tooltips for various buttons fixed and text revised",
             "- More detailed large map preview",
-            "- Seraphim icons normalized"
+            "- Seraphim icons normalized",
+            "- Both players get ready flag cleared on swap",
+            "- Removed extra space around Rerun CPU benchmark button",
+            "- Made 'Random faction' skin load the chosen faction skin (Before it would always be UEF)",
+            "- Fixed a problem preventing player colours from being updated correctly",
+            "- Prevented CPU benchmark from running once the game starts",
+            "- General performance improvements"
         }
     }
 }